modifiche ID

+++ ProgettoAlbergo/src/main/java/com/example/progettoalbergo/Model/TipologiaStanza.java
@@ -22,7 +22,7 @@ public class TipologiaStanza {
     	
     }
     
-	public TipologiaStanza(Integer id, String nome, BigDecimal prezzo, String descrizione, Integer capienza) {
+	public TipologiaStanza(Long id, String nome, BigDecimal prezzo, String descrizione, Integer capienza) {
 		setId(id);
 		setNome(nome);
 		setPrezzo(prezzo);
@@ -30,10 +30,10 @@ public class TipologiaStanza {
 		setCapienza(capienza);
 	}
 	
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getNome() {
